Move shared code into shared projects
