+++ src/FluentAssertions.Mvc.Shared/ControllerAssertions.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Diagnostics;
+
+namespace FluentAssertions.Mvc
+{
+    [DebuggerNonUserCode]
+	public class ControllerAssertions
+	{
+		public ControllerAssertions ()
+		{
+		}
+	}
+}
+
