+++ src/FluentAssertions.Mvc.Shared/FluentAssertions.Mvc.Shared.projitems
@@ -1,0 +1,34 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>0e46fc22-eb6a-431e-bfb3-47fbc4fc500f</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>FluentAssertions.Mvc.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)FailureMessages.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)ActionResultAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)AssertionsExtensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ContentResultAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ControllerAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)FailureMessages.Designer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fakes\FakeHttpContext.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fakes\FakeHttpRequest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fakes\FakeHttpResponse.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)PartialViewResultAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RedirectResultAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RedirectToRouteAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RouteCollection_Extensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RouteDataAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RouteValueDictionary_Extensions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ViewResultAssertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ViewResultBaseAssertions.cs" />
+  </ItemGroup>
+</Project>
