--- src/FluentAssertions.Mvc3/ControllerAssertions.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Diagnostics;
-
-namespace FluentAssertions.Mvc
-{
-    [DebuggerNonUserCode]
-	public class ControllerAssertions
-	{
-		public ControllerAssertions ()
-		{
-		}
-	}
-}
-
