+++ src/FluentAssertions.Mvc3/FluentAssertions.Mvc3.csproj
@@ -93,40 +93,14 @@
     <Compile Include="..\SolutionInfo.cs">
       <Link>Properties\SolutionInfo.cs</Link>
     </Compile>
-    <Compile Include="ContentResultAssertions.cs" />
-    <Compile Include="FailureMessages.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>FailureMessages.resx</DependentUpon>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="AssertionsExtensions.cs" />
-    <Compile Include="ControllerAssertions.cs" />
-    <Compile Include="ActionResultAssertions.cs" />
-    <Compile Include="Fakes\FakeHttpContext.cs" />
-    <Compile Include="Fakes\FakeHttpRequest.cs" />
-    <Compile Include="Fakes\FakeHttpResponse.cs" />
-    <Compile Include="PartialViewResultAssertions.cs" />
-    <Compile Include="RedirectResultAssertions.cs" />
-    <Compile Include="RedirectToRouteAssertions.cs" />
-    <Compile Include="RouteCollection_Extensions.cs" />
-    <Compile Include="RouteDataAssertions.cs" />
-    <Compile Include="RouteValueDictionary_Extensions.cs" />
-    <Compile Include="ViewResultAssertions.cs" />
-    <Compile Include="ViewResultBaseAssertions.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="FailureMessages.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>FailureMessages.Designer.cs</LastGenOutput>
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
+  <Import Project="..\FluentAssertions.Mvc.Shared\FluentAssertions.Mvc.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
 </Project>
