+++ src/FluentAssertions.Mvc4/FluentAssertions.Mvc4.csproj
@@ -81,69 +81,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\FluentAssertions.Mvc3\ActionResultAssertions.cs">
-      <Link>ActionResultAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\AssertionsExtensions.cs">
-      <Link>AssertionsExtensions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\ContentResultAssertions.cs">
-      <Link>ContentResultAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\ControllerAssertions.cs">
-      <Link>ControllerAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\FailureMessages.Designer.cs">
-      <Link>FailureMessages.Designer.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\Fakes\FakeHttpContext.cs">
-      <Link>Fakes\FakeHttpContext.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\Fakes\FakeHttpRequest.cs">
-      <Link>Fakes\FakeHttpRequest.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\Fakes\FakeHttpResponse.cs">
-      <Link>Fakes\FakeHttpResponse.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\PartialViewResultAssertions.cs">
-      <Link>PartialViewResultAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\RedirectResultAssertions.cs">
-      <Link>RedirectResultAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\RedirectToRouteAssertions.cs">
-      <Link>RedirectToRouteAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\RouteCollection_Extensions.cs">
-      <Link>RouteCollection_Extensions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\RouteDataAssertions.cs">
-      <Link>RouteDataAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\RouteValueDictionary_Extensions.cs">
-      <Link>RouteValueDictionary_Extensions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\ViewResultAssertions.cs">
-      <Link>ViewResultAssertions.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3\ViewResultBaseAssertions.cs">
-      <Link>ViewResultBaseAssertions.cs</Link>
-    </Compile>
     <Compile Include="..\SolutionInfo.cs">
       <Link>Properties\SolutionInfo.cs</Link>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="..\FluentAssertions.Mvc3\FailureMessages.resx">
-      <Link>FailureMessages.resx</Link>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
+  <Import Project="..\FluentAssertions.Mvc.Shared\FluentAssertions.Mvc.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
