+++ tests/FluentAssertions.Mvc3.Tests/FluentAssertions.Mvc3.Tests.csproj
@@ -84,18 +84,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ContentResultAssertions_Tests.cs" />
-    <Compile Include="FailureMessageHelper_Tests.cs" />
-    <Compile Include="Helpers\FailureMessageHelper.cs" />
-    <Compile Include="PartialViewResultAssertions_Tests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Fakes\FakeController.cs" />
-    <Compile Include="ActionResultAssertions_Tests.cs" />
-    <Compile Include="RedirectResultAssertions_Tests.cs" />
-    <Compile Include="RedirectToRoute_Tests.cs" />
-    <Compile Include="RouteDataAssertions_Tests.cs" />
-    <Compile Include="RouteValueDictionary_Extensions_Tests.cs" />
-    <Compile Include="ViewResultAssertions_Tests.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
@@ -109,5 +98,6 @@
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
+  <Import Project="..\FluentAssertions.Mvc.Tests.Shared\FluentAssertions.Mvc.Tests.Shared.projitems" Label="Shared" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
 </Project>
