+++ tests/FluentAssertions.Mvc4.Tests/FluentAssertions.Mvc4.Tests.csproj
@@ -84,39 +84,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\ActionResultAssertions_Tests.cs">
-      <Link>ActionResultAssertions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\ContentResultAssertions_Tests.cs">
-      <Link>ContentResultAssertions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\FailureMessageHelper_Tests.cs">
-      <Link>FailureMessageHelper_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\Fakes\FakeController.cs">
-      <Link>Fakes\FakeController.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\Helpers\FailureMessageHelper.cs">
-      <Link>Helpers\FailureMessageHelper.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\PartialViewResultAssertions_Tests.cs">
-      <Link>PartialViewResultAssertions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\RedirectResultAssertions_Tests.cs">
-      <Link>RedirectResultAssertions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\RedirectToRoute_Tests.cs">
-      <Link>RedirectToRoute_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\RouteDataAssertions_Tests.cs">
-      <Link>RouteDataAssertions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\RouteValueDictionary_Extensions_Tests.cs">
-      <Link>RouteValueDictionary_Extensions_Tests.cs</Link>
-    </Compile>
-    <Compile Include="..\FluentAssertions.Mvc3.Tests\ViewResultAssertions_Tests.cs">
-      <Link>ViewResultAssertions_Tests.cs</Link>
-    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -130,6 +97,7 @@
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
+  <Import Project="..\FluentAssertions.Mvc.Tests.Shared\FluentAssertions.Mvc.Tests.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
